Atmosphere switching - release v2.1.0 (#303)
* Atmosphere wrapper introduced for switching between extensions without re-installing X-PSI

* Adding the wrapper file missing from the previous commit

* Making basic numerical test run working as well.

* Atmosphere options passed as a string from the user instead of an integer.

* Function renamings and beaming moved the wrapper.

Beaming modification is now mooved to the wrapper function, and it can be used for many atmosphere extensions (instead of having it as a separate extension). Setting it on or off is now based on an additional input argument.

* Switching between elsewhere atmosphere extensions implemented.

* Error handling if atmosphere data not provided.

Exception raised now in Python if atmosphere data were not provided, even though using numerical atmosphere extension.

* Passing atmosphere also for cede=True.

* Surface radiation field tools upgraded.

Tutorial and the related intensity functions in X-PSI upgraded for the atmosphere switching without re-installing X-PSI.

* Atmosphere options implented also to Everywhere and integrator.py.

* Added option for user extension.

An option added for user to provide their own atmosphere extension by replacing some files. In addition, atmosphere options included now also for imaging in Photosphere and pixelmesh.

* Installation+documentation updates, and cleaning.

* Version numbers and setup description updated.

* Doc updates 1

* Now unnecessary archives removed.

* Doc updates 2

Tutorials re-run and comments updated.

* Doc updates 3

More tutorials re-run and updated.

* Module generator tutorial updated.

* CHANGELOG and small doc updates.

* Adding beam_opt value checks

in everywhere and in `HotRegion.py` and in `Everywhere.py`

* Adding elsewhere_user in the wrapper

* required versions of cython

for CI tests

* Update basic_environment.yml

* cython version requirements

* Notebook Surface Radiation Field Tool completed

with extra info to create a CustomPhotosphere with atmosphere extension 'Num4D'

* Meaning of old scripts clarified.

* small cleaning

* warning if loading atmosphere data for default user elsewhere BB

* Added sanity checks in hot_wrapper and elsewhere_wrapper

Added checks to default to BB in case the `atmos_extension` values are different than the expected 1, 2, 3.  It should not happen, but in case it does, the crash might be difficult to understand.

* Small version date and number fixes.

---------

Co-authored-by: sguillot <sebastien.guillot@irap.omp.eu>