Precisions for parameters in corner plots
Following Basdorsman's suggestion in issue #393, we develop here a new structure that allows to choose each precision for the parameters loaded from the runs and showed in the corner plots.
We extended the Runs tool so it can assign chosen precisions for the  parameters, using a distribution of integers and/or None objects, which, in the latter case makes the corner plots to choose the original default precision for the involved parameters.