Update Signal.py
You may face another issue when you activate pileup for some instrument files which have their response trimmed. In that case, you store a copy of the original instrument response for pileup checking. However this copy wont be trimmed as the original one when you determine the signal. Consequently, the estimated likelihood will be, with the pileup module activated, in extreme scales (-1e90) for any sample. 
WIth this new update, the instrument response stored for pileup checking will be trimmed as the one used for the signal.