Energy Integration with discontinuous intrumental responses
In this draft, based on a mixing between pull requests #503 and #588, a different energy integrator has been tested, in the case of signals presenting bad quality data among good ones. As X-PSI is initially based on integration on continuous bins, the use of instrumental responses with "holes" is not adapted in the initial configuration, returning the response error related to the lack of positive values or the assertion errors related to the energies and channel edges.
So a new version of the energy integrator is proposed here, with the edges designed into 2 arrays: one with the lower bounds and one with the upper bounds for each energy bin.
Another extension is also given to enable X-PSI to use linear interpolation.